+++ src/reducers/moodReducers.js
@@ -6,19 +6,19 @@ import {
 } from '../actions/moodActions';
 
 export default function reducer(state, action) {
-  switch(action.type) {
-    case DRINK_COFFEE: {
+  switch (action.type) {
+    case DRINK_COFFEE:
       return { ...state, coffees: state.coffees + 1 };
-    }
-    case EAT_SNACK: {
+
+    case EAT_SNACK:
       return { ...state, snacks: state.snacks + 1 };
-    }
-    case TAKE_NAP: {
+
+    case TAKE_NAP:
       return { ...state, naps: state.naps + 1 };
-    }
-    case STUDY: {
-      return { ...state, study: state.study + 1 };
-    }
+
+    case STUDY:
+      return { ...state, studies: state.studies + 1 };
+
     default:
       return state;
   }
